+++ src/main/features/event-management/controller/steps-to-continue.ts
@@ -2,6 +2,7 @@ import * as express from 'express'
 import { LoggTracer } from '@common/logtracer/tracer'
 import { TokenDecoder } from '@common/tokendecoder/tokendecoder'
 import * as data from '../../../resources/content/event-management/steps-to-continue.json'
+import { TenderApi } from './../../../common/util/fetch/procurementService/TenderApiInstance';
 
 /**
  * 
@@ -56,6 +57,21 @@ export const POST_STEPS_TO_CONTINUE = async (req: express.Request, res: express.
   try {
    
     const  rfi_next_steps  =  req.body.rfi_next_steps_to_continue;
+    
+    let baseUrl = `/tenders/projects/${req.session.projectId}/events`;
+    let body = {
+      "name": "Further Competition Event",
+      "eventType": "FCA"
+    }
+    const { data } = await TenderApi.Instance(SESSION_ID).post(baseUrl, body);
+    if(data != null && data !=undefined)
+    {
+      req.session['eventId'] = data.id;
+      req.session.procurements[0]['eventId'] = data.id;
+      req.session.procurements[0]['eventType'] = data.eventType;
+      req.session.procurements[0]['started'] = false;
+    }
+
     if (rfi_next_steps) {
       switch (rfi_next_steps) {
         case '[DA]':
