+++ src/main/features/event-management/views/standstillPeriodDecision.njk
@@ -60,11 +60,11 @@
             {# <strong class="govuk-tag app-task-list__tag govuk-tag--grey" id="suppliers-status">{{status}}</strong> #}
           </h2>
           <p class="govuk-body">
-            {% if(showCloseProject) %}
+            {#{% if(showCloseProject) %}#}
             <a class="govuk-link" href="/rfi/nextsteps">Close Event</a>
-            {% else %}
+           {# {% else %}
             <a class="govuk-link" href="/event/management?id={{ eventId }}">Close Event</a>
-            {% endif %}
+            {% endif %} #}
           </p>
         </div>
       </div>
